draftivist_api/requirements.txt
boto3==1.16
Django==3.0
django-cors-headers==3.4
django-storages==1.10
djangorestframework==3.11
psycopg2==2.8.6
python-dotenv==0.15

# TODO: Prod only
gunicorn==20.0
